LPS-107004 Apply renames
